requirements.txt
adjustText==1.1.1
apex==0.9.10dev
dlib==19.24.4
facenet_pytorch==2.5.2
ignite==1.1.0
kornia==0.7.2
lmdb==1.4.1
matplotlib==3.8.2
ml_collections==0.1.1
monai==1.3.1
numpy==1.26.4
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
pytorch_fid==0.2.1
PyYAML==6.0.1
scipy==1.13.1
seaborn==0.13.2
tensorboardX==2.6.2.2
timm==0.6.13
tqdm==4.66.1
wandb==0.16.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cu118
torchvision==0.15.2+cu118
